+++ tests/builtin_open_test.py
@@ -41,6 +41,8 @@ class TestSecureCodingStandardChecker(pylint.testutils.CheckerTestCase):
         with_nodes = nodes[5:]
 
         with self.assertNoMessages():
+            self.checker.set_os_open_mode('True')
+            assert self.checker._prefer_os_open
             for idx, node in enumerate(call_nodes):
                 self.checker.visit_call(node)
             for idx, node in enumerate(with_nodes):
@@ -70,13 +72,27 @@ class TestSecureCodingStandardChecker(pylint.testutils.CheckerTestCase):
     )
 
     @pytest.mark.parametrize('s', _calls_not_ok)
-    def test_builtin_open_call(self, s):
+    @pytest.mark.parametrize('os_open_mode', (False, True))
+    def test_builtin_open_call(self, s, os_open_mode):
         node = astroid.extract_node(s + ' #@')
-        with self.assertAddsMessages(pylint.testutils.Message(msg_id='replace-builtin-open', node=node)):
-            self.checker.visit_call(node)
+        self.checker.set_os_open_mode(str(os_open_mode))
+        assert self.checker._prefer_os_open == os_open_mode
+        if os_open_mode:
+            with self.assertAddsMessages(pylint.testutils.Message(msg_id='replace-builtin-open', node=node)):
+                self.checker.visit_call(node)
+        else:
+            with self.assertNoMessages():
+                self.checker.visit_call(node)
 
     @pytest.mark.parametrize('s', ('with ' + s + ' as fd: fd.read()' for s in _calls_not_ok))
-    def test_builtin_open_with(self, s):
+    @pytest.mark.parametrize('os_open_mode', (False, True))
+    def test_builtin_open_with(self, s, os_open_mode):
         node = astroid.extract_node(s + ' #@')
-        with self.assertAddsMessages(pylint.testutils.Message(msg_id='replace-builtin-open', node=node)):
-            self.checker.visit_with(node)
+        self.checker.set_os_open_mode(str(os_open_mode))
+        assert self.checker._prefer_os_open == os_open_mode
+        if os_open_mode:
+            with self.assertAddsMessages(pylint.testutils.Message(msg_id='replace-builtin-open', node=node)):
+                self.checker.visit_with(node)
+        else:
+            with self.assertNoMessages():
+                self.checker.visit_with(node)